+++ CHANGELOG.md
@@ -32,6 +32,7 @@
    (i.e. only numbers such as 1899 or 1923),
    we assumed it indeed is an ID and not a search query (#1527).
  - TMDb Concerts: We didn't always properly save the currently used language.
+ - HotMovies: Movie search works again
 
 ### Changes